requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
Pillow==6.2.0
postgres==3.0.0
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pyTelegramBotAPI==3.6.6
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
python-decouple = "*"
django-filter= "*"
